requirements.txt
python-binance
tweepy
python-dotenv
pandas
numpy
TA-Lib
vaderSentiment
matplotlib
schedule
